Collapse expand
Simplistic control to collapse expand the whole structure.

The idea is pretty simple. For a given structure, the display can either
be :all or :hide. An :all / :hide value will propagate to the subtree of
the structure. For indexed structure such as Map/Objects Vector/Arrays
the structure can be replicated such that in front of each index you
have the :all / :hide value. This value will be passed alongside the
path to each display-data call but some types may not take this into
account, such as strings, numbers, boolean (every none compound type).

This commit only adds this behaviour at the top level of the data
structure.